% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datiComuniItalia}
\alias{datiComuniItalia}
\title{Dati comuni italiani forniti dall'ISTAT(2022)}
\format{
dataframe con 7904 righe e 3 variabili:
\describe{
\item{regione: denominazione regione di appartenenza del comune}
\item{provincia: denominazione provincia di appartenenza del comune}
\item{comune: denominazione comune}
}
\usage{
datiComuniItalia
}
\description{
Dati comuni italiani forniti dall'ISTAT(2022)
}
\keyword{datasets}
